List of updates:
- reconfigure k8s config to enable bootstrapping deployment (for offline development using minikube)
- Change mongodb from deployment to statefulset kind to enable scaling.
- Add least privileged audituser (app specific user) for mongo.
- Change initial data insertion & migration support using flyway.
- Add tracking number controller & related unit tests.

+++ src/main/java/com/tele/microservice/entity/OrderRecord.java
@@ -5,6 +5,7 @@ import jakarta.validation.constraints.*;
 import lombok.*;
 import org.hibernate.annotations.CreationTimestamp;
 
+import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
@@ -41,7 +42,7 @@ public class OrderRecord {
     @Column(nullable = false, precision = 6, scale = 3)
     @Positive
     @DecimalMin("0.001")
-    private Double weight;  // in kilograms, up to 3 decimal places
+    private BigDecimal weight;  // in kilograms, up to 3 decimal places
 
     @Column(name = "created_at", nullable = false, updatable = false)
     @CreationTimestamp
